refactored api call to separate util component. Used knex to create, migrate, and seed fetched data to mySQL database
