+++ server.js
@@ -1,12 +1,15 @@
 const express = require("express");
-const dotenv = require("dotenv");
-const billsRoutes = require("./src/routes/billsRoutes");
-const openaiRoutes = require("./src/routes/openaiRoutes");
-
-dotenv.config(); // Load environment variables from .env
-
 const app = express();
-const port = process.env.PORT || 3000;
+const cors = require("cors");
+const dotenv = require("dotenv");
+dotenv.config(); // Load environment variables from .env
+const port = process.env.PORT || 8000;
+const billsRoutes = require("./src/routes/bills-Routes");
+const openaiRoutes = require("./src/routes/openai-Routes");
+
+// express and cors middleware
+app.use(express.json());
+app.use(cors());
 
 // Use billsRoutes for bills-related routes
 app.use("/api", billsRoutes);
